LSF corrections and new tests
