+++ smartsim/launcher/pbs/pbsLauncher.py
@@ -119,8 +119,6 @@ class PBSLauncher(WLMLauncher):
             step_id = self._get_pbs_step_id(step)
         self.step_mapping.add(step.name, step_id, task_id, step.managed)
 
-        time.sleep(5)
-
         return step_id
 
     def stop(self, step_name):
